Move out some classes into their own files
