+++ mapcache/src/main/java/mil/nga/mapcache/GeoPackageMapFragment.java
@@ -17,7 +17,6 @@ import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Handler;
 import android.os.Looper;
 import android.os.Vibrator;
 import android.preference.PreferenceManager;
@@ -110,7 +109,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
@@ -123,7 +121,6 @@ import mil.nga.geopackage.db.GeoPackageDataType;
 import mil.nga.geopackage.extension.nga.link.FeatureTileTableLinker;
 import mil.nga.geopackage.extension.nga.scale.TileScaling;
 import mil.nga.geopackage.extension.nga.scale.TileTableScaling;
-import mil.nga.geopackage.extension.nga.style.FeatureStyle;
 import mil.nga.geopackage.extension.schema.SchemaExtension;
 import mil.nga.geopackage.extension.schema.columns.DataColumns;
 import mil.nga.geopackage.extension.schema.columns.DataColumnsDao;
@@ -141,17 +138,10 @@ import mil.nga.geopackage.geom.GeoPackageGeometryData;
 import mil.nga.geopackage.map.MapUtils;
 import mil.nga.geopackage.map.features.FeatureInfoBuilder;
 import mil.nga.geopackage.map.features.StyleCache;
-import mil.nga.geopackage.map.geom.FeatureShapes;
 import mil.nga.geopackage.map.geom.GoogleMapShape;
 import mil.nga.geopackage.map.geom.GoogleMapShapeConverter;
 import mil.nga.geopackage.map.geom.GoogleMapShapeMarkers;
 import mil.nga.geopackage.map.geom.GoogleMapShapeType;
-import mil.nga.geopackage.map.geom.MultiLatLng;
-import mil.nga.geopackage.map.geom.MultiMarker;
-import mil.nga.geopackage.map.geom.MultiPolygon;
-import mil.nga.geopackage.map.geom.MultiPolygonOptions;
-import mil.nga.geopackage.map.geom.MultiPolyline;
-import mil.nga.geopackage.map.geom.MultiPolylineOptions;
 import mil.nga.geopackage.map.geom.PolygonHoleMarkers;
 import mil.nga.geopackage.map.geom.ShapeMarkers;
 import mil.nga.geopackage.map.geom.ShapeWithChildrenMarkers;
@@ -193,7 +183,6 @@ import mil.nga.mapcache.repository.GeoPackageModifier;
 import mil.nga.mapcache.sensors.SensorHandler;
 import mil.nga.mapcache.utils.ProjUtils;
 import mil.nga.mapcache.utils.SwipeController;
-import mil.nga.mapcache.utils.ThreadUtils;
 import mil.nga.mapcache.utils.ViewAnimation;
 import mil.nga.mapcache.view.GeoPackageAdapter;
 import mil.nga.mapcache.view.detail.DetailActionUtil;
@@ -331,11 +320,6 @@ public class GeoPackageMapFragment extends Fragment implements
     private final Lock updateLock = new ReentrantLock();
 
     /**
-     * Mapping of open GeoPackage feature DAOs
-     */
-    private final Map<String, Map<String, FeatureDao>> featureDaos = new HashMap<>();
-
-    /**
      * Vibrator
      */
     private Vibrator vibrator;
@@ -349,11 +333,6 @@ public class GeoPackageMapFragment extends Fragment implements
      * Bounding box mode
      */
     private boolean boundingBoxMode = false;
-
-    /**
-     * Edit features mode
-     */
-    private boolean editFeaturesMode = false;
 
     /**
      * Bounding box starting corner
@@ -386,44 +365,9 @@ public class GeoPackageMapFragment extends Fragment implements
     private MenuItem editFeaturesMenuItem;
 
     /**
-     * Edit features database
-     */
-    private String editFeaturesDatabase;
-
-    /**
-     * Edit features table
-     */
-    private String editFeaturesTable;
-
-    /**
-     * Feature shapes
-     */
-    private final FeatureShapes featureShapes = new FeatureShapes();
-
-    /**
      * Current zoom level
      */
     private int currentZoom = -1;
-
-    /**
-     * Flag indicating if the initial zoom is still needed
-     */
-    private boolean needsInitialZoom = true;
-
-    /**
-     * Mapping between marker ids and the feature ids
-     */
-    private final Map<String, Long> editFeatureIds = new HashMap<>();
-
-    /**
-     * Mapping between marker ids and the features
-     */
-    private final Map<String, MarkerFeature> markerIds = new HashMap<>();
-
-    /**
-     * Mapping between marker ids and feature objects
-     */
-    private final Map<String, GoogleMapShape> editFeatureObjects = new HashMap<>();
 
     /**
      * Edit points type
@@ -528,11 +472,6 @@ public class GeoPackageMapFragment extends Fragment implements
      * Edit clear button
      */
     private ImageButton editClearPolygonHolesButton;
-
-    /**
-     * Lock for concurrently updating the features bounding box
-     */
-    private final Lock featuresBoundingBoxLock = new ReentrantLock();
 
     /**
      * List of Feature Overlay Queries for querying tile overlay clicks
@@ -1380,7 +1319,7 @@ public class GeoPackageMapFragment extends Fragment implements
 
             // Set text for edit features mode
             MenuItem editFeaturesItem = pm.getMenu().findItem(R.id.features);
-            if (editFeaturesMode) {
+            if (model.isEditFeaturesMode()) {
                 editFeaturesItem.setTitle("Stop editing");
             } else {
                 editFeaturesItem.setTitle("Edit Features");
@@ -1413,7 +1352,7 @@ public class GeoPackageMapFragment extends Fragment implements
                     return true;
                 } else if (item.getItemId() == R.id.features) {
                     editFeaturesMenuItem = item;
-                    if (!editFeaturesMode) {
+                    if (!model.isEditFeaturesMode()) {
                         selectEditFeatures();
                     } else {
                         resetEditFeatures();
@@ -1424,7 +1363,7 @@ public class GeoPackageMapFragment extends Fragment implements
                     boundingBoxMenuItem = item;
                     if (!boundingBoxMode) {
 
-                        if (editFeaturesMode) {
+                        if (model.isEditFeaturesMode()) {
                             resetEditFeatures();
                             updateInBackground(false, true);
                         }
@@ -2377,17 +2316,17 @@ public class GeoPackageMapFragment extends Fragment implements
     public void onCameraIdle() {
 
         // If visible & not editing a shape, update the feature shapes for the current map view region
-        if (visible && (!editFeaturesMode || editFeatureType == null || (editPoints.isEmpty() && editFeatureMarker == null))) {
+        if (visible && (!model.isEditFeaturesMode() || editFeatureType == null || (editPoints.isEmpty() && editFeatureMarker == null))) {
 
             int previousZoom = currentZoom;
             int zoom = (int) MapUtils.getCurrentZoom(map);
             currentZoom = zoom;
             if (zoom != previousZoom) {
                 // Zoom level changed, remove all feature shapes except for markers
-                featureShapes.removeShapesExcluding(GoogleMapShapeType.MARKER, GoogleMapShapeType.MULTI_MARKER);
+                model.getFeatureShapes().removeShapesExcluding(GoogleMapShapeType.MARKER, GoogleMapShapeType.MULTI_MARKER);
             } else {
                 // Remove shapes no longer visible on the map view
-                featureShapes.removeShapesNotWithinMap(map);
+                model.getFeatureShapes().removeShapesNotWithinMap(map);
             }
 
             BoundingBox mapViewBoundingBox = MapUtils.getBoundingBox(map);
@@ -2399,7 +2338,7 @@ public class GeoPackageMapFragment extends Fragment implements
                 if (updateFeaturesTask != null) {
                     updateFeaturesTask.cancel(false);
                 }
-                updateFeaturesTask = new MapFeaturesUpdateTask();
+                updateFeaturesTask = new MapFeaturesUpdateTask(getContext(), map, model, geoPackageViewModel, zoomer);
                 updateFeaturesTask.execute(false, maxFeatures, mapViewBoundingBox, toleranceDistance, true);
             } finally {
                 updateLock.unlock();
@@ -2611,11 +2550,11 @@ public class GeoPackageMapFragment extends Fragment implements
                 case EDIT_FEATURE:
                     editFeatureMarker = tempEditFeatureMarker;
                     tempEditFeatureMarker = null;
-                    Long featureId = editFeatureIds.get(editFeatureMarker.getId());
+                    Long featureId = model.getEditFeatureIds().get(editFeatureMarker.getId());
                     if (featureId != null) {
-                        final GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(editFeaturesDatabase);
+                        final GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(model.getEditFeaturesDatabase());
                         final FeatureDao featureDao = geoPackage
-                                .getFeatureDao(editFeaturesTable);
+                                .getFeatureDao(model.getEditFeaturesTable());
                         final FeatureRow featureRow = featureDao
                                 .queryForIdRow(featureId);
                         Geometry geometry = featureRow.getGeometry().getGeometry();
@@ -2624,7 +2563,7 @@ public class GeoPackageMapFragment extends Fragment implements
                         GoogleMapShape shape = converter.toShape(geometry);
 
                         editFeatureMarker.remove();
-                        GoogleMapShape featureObject = editFeatureObjects
+                        GoogleMapShape featureObject = model.getEditFeatureObjects()
                                 .remove(editFeatureMarker.getId());
                         if (featureObject != null) {
                             featureObject.remove();
@@ -2650,11 +2589,11 @@ public class GeoPackageMapFragment extends Fragment implements
      */
     private void addEditableShapeBack() {
 
-        Long featureId = editFeatureIds.get(editFeatureMarker.getId());
+        Long featureId = model.getEditFeatureIds().get(editFeatureMarker.getId());
         if (featureId != null) {
-            final GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(editFeaturesDatabase);
+            final GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(model.getEditFeaturesDatabase());
             final FeatureDao featureDao = geoPackage
-                    .getFeatureDao(editFeaturesTable);
+                    .getFeatureDao(model.getEditFeaturesTable());
             final FeatureRow featureRow = featureDao.queryForIdRow(featureId);
             GeoPackageGeometryData geomData = featureRow.getGeometry();
             if (geomData != null) {
@@ -2664,10 +2603,17 @@ public class GeoPackageMapFragment extends Fragment implements
                             featureDao.getProjection());
                     GoogleMapShape shape = converter.toShape(geometry);
                     StyleCache styleCache = new StyleCache(geoPackage, getResources().getDisplayMetrics().density);
-                    prepareShapeOptions(shape, styleCache, featureRow, true, true);
+                    ShapeHelper.getInstance().prepareShapeOptions(
+                            shape,
+                            styleCache,
+                            featureRow,
+                            true,
+                            true,
+                            getContext());
                     GoogleMapShape mapShape = GoogleMapShapeConverter
                             .addShapeToMap(map, shape);
-                    addEditableShape(featureId, mapShape);
+                    ShapeHelper.getInstance().addEditableShape(
+                            getContext(), map, model, featureId, mapShape);
                     styleCache.clear();
                 }
             }
@@ -2738,10 +2684,10 @@ public class GeoPackageMapFragment extends Fragment implements
 
         boolean changesMade = false;
 
-        GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(editFeaturesDatabase);
+        GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(model.getEditFeaturesDatabase());
         EditType tempEditFeatureType = editFeatureType;
         try {
-            FeatureDao featureDao = geoPackage.getFeatureDao(editFeaturesTable);
+            FeatureDao featureDao = geoPackage.getFeatureDao(model.getEditFeaturesTable());
             long srsId = featureDao.getGeometryColumns().getSrsId();
             FeatureIndexManager indexer = new FeatureIndexManager(getActivity(), geoPackage, featureDao);
             List<FeatureIndexType> indexedTypes = indexer.getIndexedTypes();
@@ -2808,7 +2754,7 @@ public class GeoPackageMapFragment extends Fragment implements
 
                 case EDIT_FEATURE:
                     editFeatureType = null;
-                    Long featureId = editFeatureIds.get(editFeatureMarker.getId());
+                    Long featureId = model.getEditFeatureIds().get(editFeatureMarker.getId());
 
                     if (featureId != null) {
                         Geometry geometry = converter.toGeometry(editFeatureShape
@@ -2848,7 +2794,7 @@ public class GeoPackageMapFragment extends Fragment implements
                         .showMessage(
                                 getActivity(),
                                 getString(R.string.edit_features_save_label)
-                                        + " " + editFeaturesTable,
+                                        + " " + model.getEditFeaturesTable(),
                                 "GeoPackage contains unsupported SQLite function, module, or trigger for writing: " + e.getMessage());
             } else {
                 GeoPackageUtils.showMessage(getActivity(),
@@ -2954,7 +2900,7 @@ public class GeoPackageMapFragment extends Fragment implements
             handled = true;
         } else if (item.getItemId() == R.id.map_features) {
             editFeaturesMenuItem = item;
-            if (!editFeaturesMode) {
+            if (!model.isEditFeaturesMode()) {
                 selectEditFeatures();
             } else {
                 resetEditFeatures();
@@ -2965,7 +2911,7 @@ public class GeoPackageMapFragment extends Fragment implements
             boundingBoxMenuItem = item;
             if (!boundingBoxMode) {
 
-                if (editFeaturesMode) {
+                if (model.isEditFeaturesMode()) {
                     resetEditFeatures();
                     updateInBackground(false, true);
                 }
@@ -2996,10 +2942,10 @@ public class GeoPackageMapFragment extends Fragment implements
                 resetBoundingBox();
             }
 
-            editFeaturesDatabase = geoPackage;
-            editFeaturesTable = layer;
+            model.setEditFeaturesDatabase(geoPackage);
+            model.setEditFeaturesTable(layer);
 
-            editFeaturesMode = true;
+            model.setEditFeaturesMode(true);
             editFeaturesView.setVisibility(View.VISIBLE);
 
 
@@ -3042,12 +2988,10 @@ public class GeoPackageMapFragment extends Fragment implements
                                 resetBoundingBox();
                             }
 
-                            editFeaturesDatabase = geoPackageInput
-                                    .getSelectedItem().toString();
-                            editFeaturesTable = featuresInput.getSelectedItem()
-                                    .toString();
+                            model.setEditFeaturesDatabase(geoPackageInput.getSelectedItem().toString());
+                            model.setEditFeaturesTable(featuresInput.getSelectedItem().toString());
 
-                            editFeaturesMode = true;
+                            model.setEditFeaturesMode(true);
                             editFeaturesView.setVisibility(View.VISIBLE);
                             editFeaturesMenuItem
                                     .setIcon(R.drawable.ic_features_active);
@@ -3095,12 +3039,12 @@ public class GeoPackageMapFragment extends Fragment implements
      * Reset the edit features state
      */
     private void resetEditFeatures() {
-        editFeaturesMode = false;
+        model.setEditFeaturesMode(false);
         editFeaturesView.setVisibility(View.INVISIBLE);
-        editFeaturesDatabase = null;
-        editFeaturesTable = null;
-        editFeatureIds.clear();
-        editFeatureObjects.clear();
+        model.setEditFeaturesDatabase(null);
+        model.setEditFeaturesTable(null);
+        model.getEditFeatureIds().clear();
+        model.getEditFeatureObjects().clear();
         editFeatureShape = null;
         editFeatureShapeMarkers = null;
         editFeatureMarker = null;
@@ -3352,7 +3296,7 @@ public class GeoPackageMapFragment extends Fragment implements
     private void updateInBackground(boolean zoom, boolean filter) {
         if (getActivity() != null) {
             getActivity().runOnUiThread(() -> map.clear());
-            featureDaos.clear();
+            model.getFeatureDaos().clear();
             basemapApplier.clear();
 
             if (zoom) {
@@ -3363,8 +3307,8 @@ public class GeoPackageMapFragment extends Fragment implements
             model.setTilesBoundingBox(null);
             model.setFeatureOverlayTiles(false);
             featureOverlayQueries.clear();
-            featureShapes.clear();
-            markerIds.clear();
+            model.getFeatureShapes().clear();
+            model.getMarkerIds().clear();
             int maxFeatures = getMaxFeatures();
 
             getActivity().runOnUiThread(() -> {
@@ -3458,7 +3402,7 @@ public class GeoPackageMapFragment extends Fragment implements
 
                         if (!featureTableDaos.isEmpty()) {
                             Map<String, FeatureDao> databaseFeatureDaos = new HashMap<>();
-                            featureDaos.put(database.getDatabase(), databaseFeatureDaos);
+                            model.getFeatureDaos().put(database.getDatabase(), databaseFeatureDaos);
                             for (String featureTable : featureTableDaos) {
 
                                 if (task.isCancelled()) {
@@ -3513,7 +3457,7 @@ public class GeoPackageMapFragment extends Fragment implements
                     if (updateFeaturesTask != null) {
                         updateFeaturesTask.cancel(false);
                     }
-                    updateFeaturesTask = new MapFeaturesUpdateTask();
+                    updateFeaturesTask = new MapFeaturesUpdateTask(getContext(), map, model, geoPackageViewModel, zoomer);
                     updateFeaturesTask.execute(zoom, maxFeatures, mapViewBoundingBox, toleranceDistance, filter);
                 } finally {
                     updateLock.unlock();
@@ -3522,797 +3466,6 @@ public class GeoPackageMapFragment extends Fragment implements
 
         }
 
-    }
-
-    /**
-     * Update the map features in the background
-     */
-    private class MapFeaturesUpdateTask extends AsyncTask<Object, Object, Integer> {
-
-        /**
-         * Zoom after update flag
-         */
-        private boolean zoom;
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        protected Integer doInBackground(Object... params) {
-            zoom = (Boolean) params[0];
-            int maxFeatures = (Integer) params[1];
-            BoundingBox mapViewBoundingBox = (BoundingBox) params[2];
-            double toleranceDistance = (Double) params[3];
-            boolean filter = (Boolean) params[4];
-            return addFeatures(this, maxFeatures, mapViewBoundingBox, toleranceDistance, filter);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        protected void onProgressUpdate(Object... shapeUpdate) {
-
-            long featureId = (Long) shapeUpdate[0];
-            String database = (String) shapeUpdate[1];
-            String tableName = (String) shapeUpdate[2];
-            GoogleMapShape shape = (GoogleMapShape) shapeUpdate[3];
-
-            synchronized (featureShapes) {
-
-                if (!featureShapes.exists(featureId, database, tableName)) {
-
-                    GoogleMapShape mapShape = GoogleMapShapeConverter.addShapeToMap(
-                            map, shape);
-
-                    if (editFeaturesMode) {
-                        Marker marker = addEditableShape(featureId, mapShape);
-                        if (marker != null) {
-                            GoogleMapShape mapPointShape = new GoogleMapShape(GeometryType.POINT, GoogleMapShapeType.MARKER, marker);
-                            featureShapes.addMapMetadataShape(mapPointShape, featureId, database, tableName);
-                        }
-                    } else {
-                        addMarkerShape(featureId, database, tableName, mapShape);
-                    }
-                    featureShapes.addMapShape(mapShape, featureId, database, tableName);
-                }
-            }
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        protected void onPostExecute(Integer count) {
-
-            if (needsInitialZoom || zoom) {
-                zoomer.zoomToActive(true);
-                needsInitialZoom = false;
-            }
-        }
-
-        /**
-         * Add a shape to the map
-         *
-         * @param featureId The id of the feature.
-         * @param database  The name of the geopackage.
-         * @param tableName The name of the layer.
-         * @param shape     The type of shape to add.
-         */
-        public void addToMap(long featureId, String database, String tableName, GoogleMapShape shape) {
-            publishProgress(featureId, database, tableName, shape);
-        }
-
-    }
-
-    /**
-     * Add features to the map
-     *
-     * @param task               update features task
-     * @param maxFeatures        max features
-     * @param mapViewBoundingBox map view bounding box
-     * @param toleranceDistance  tolerance distance
-     * @param filter             filter
-     * @return feature count
-     */
-    private int addFeatures(MapFeaturesUpdateTask task, final int maxFeatures, BoundingBox mapViewBoundingBox, double toleranceDistance, boolean filter) {
-
-        AtomicInteger count = new AtomicInteger();
-
-        Map<String, List<String>> featureTables = new HashMap<>();
-        if (editFeaturesMode) {
-            List<String> databaseFeatures = new ArrayList<>();
-            databaseFeatures.add(editFeaturesTable);
-            featureTables.put(editFeaturesDatabase, databaseFeatures);
-            GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(editFeaturesDatabase);
-            Map<String, FeatureDao> databaseFeatureDaos = featureDaos.get(editFeaturesDatabase);
-            if (databaseFeatureDaos == null) {
-                databaseFeatureDaos = new HashMap<>();
-                featureDaos.put(editFeaturesDatabase, databaseFeatureDaos);
-            }
-            FeatureDao featureDao = databaseFeatureDaos.get(editFeaturesTable);
-            if (featureDao == null) {
-                featureDao = geoPackage.getFeatureDao(editFeaturesTable);
-                databaseFeatureDaos.put(editFeaturesTable, featureDao);
-            }
-        } else {
-            for (GeoPackageDatabase database : model.getActive().getDatabases()) {
-                if (!database.getFeatures().isEmpty()) {
-                    List<String> databaseFeatures = new ArrayList<>();
-                    featureTables.put(database.getDatabase(),
-                            databaseFeatures);
-                    for (GeoPackageTable features : database.getFeatures()) {
-                        databaseFeatures.add(features.getName());
-                    }
-                }
-            }
-        }
-
-        for (Map.Entry<String, List<String>> databaseFeaturesEntry : featureTables
-                .entrySet()) {
-
-            if (count.get() >= maxFeatures) {
-                break;
-            }
-
-            String databaseName = databaseFeaturesEntry.getKey();
-
-            List<String> databaseFeatures = databaseFeaturesEntry.getValue();
-            Map<String, FeatureDao> databaseFeatureDaos = featureDaos.get(databaseName);
-
-            if (databaseFeatureDaos != null) {
-
-                GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(databaseName);
-                StyleCache styleCache = new StyleCache(geoPackage, getResources().getDisplayMetrics().density);
-
-                for (String features : databaseFeatures) {
-
-                    if (databaseFeatureDaos.containsKey(features)) {
-
-                        displayFeatures(task,
-                                geoPackage, styleCache, features, count,
-                                maxFeatures, editFeaturesMode, mapViewBoundingBox, toleranceDistance, filter);
-                        if (task.isCancelled() || count.get() >= maxFeatures) {
-                            break;
-                        }
-                    }
-                }
-
-                styleCache.clear();
-            }
-
-            if (task.isCancelled()) {
-                break;
-            }
-        }
-
-        return Math.min(count.get(), maxFeatures);
-    }
-
-    /**
-     * Display features
-     *
-     * @param task               The update task.
-     * @param geoPackage         The geopackage to display.
-     * @param styleCache         the style cache.
-     * @param features           The features.
-     * @param count              The number of features.
-     * @param maxFeatures        The maximum number of features the map will display.
-     * @param editable           True if its editable.
-     * @param mapViewBoundingBox The views bounding box.
-     * @param toleranceDistance  Used to simplify geometries for performance.
-     * @param filter             True if features should be filtered.
-     */
-    private void displayFeatures(MapFeaturesUpdateTask task, GeoPackage geoPackage, StyleCache styleCache, String features,
-                                 AtomicInteger count, final int maxFeatures, final boolean editable,
-                                 BoundingBox mapViewBoundingBox, double toleranceDistance, boolean filter) {
-
-        // Get the GeoPackage and feature DAO
-        String database = geoPackage.getName();
-        Map<String, FeatureDao> dataAccessObjects = featureDaos.get(database);
-        if (dataAccessObjects != null) {
-            FeatureDao featureDao = dataAccessObjects.get(features);
-            if (featureDao != null) {
-                GoogleMapShapeConverter converter = new GoogleMapShapeConverter(featureDao.getProjection());
-
-                converter.setSimplifyTolerance(toleranceDistance);
-
-                if (!styleCache.getFeatureStyleExtension().has(features)) {
-                    styleCache = null;
-                }
-
-                count.getAndAdd(featureShapes.getFeatureIdsCount(database, features));
-
-                if (!task.isCancelled() && count.get() < maxFeatures) {
-
-                    mil.nga.proj.Projection mapViewProjection = ProjectionFactory.getProjection(ProjectionConstants.EPSG_WORLD_GEODETIC_SYSTEM);
-
-                    String[] columns = featureDao.getIdAndGeometryColumnNames();
-
-                    FeatureIndexManager indexer = new FeatureIndexManager(getActivity(), geoPackage, featureDao);
-                    if (filter && indexer.isIndexed()) {
-
-                        FeatureIndexResults indexResults = indexer.query(columns, mapViewBoundingBox, mapViewProjection);
-                        BoundingBox complementary = mapViewBoundingBox.complementaryWgs84();
-                        if (complementary != null) {
-                            FeatureIndexResults indexResults2 = indexer.query(columns, complementary, mapViewProjection);
-                            indexResults = new MultipleFeatureIndexResults(indexResults, indexResults2);
-                        }
-
-                        processFeatureIndexResults(task, indexResults, database, featureDao, converter, styleCache,
-                                count, maxFeatures, editable, filter);
-
-                    } else {
-
-                        BoundingBox filterBoundingBox = null;
-                        double filterMaxLongitude = 0;
-
-                        if (filter) {
-                            mil.nga.proj.Projection featureProjection = featureDao.getProjection();
-                            ProjectionTransform projectionTransform = mapViewProjection.getTransformation(featureProjection);
-                            BoundingBox boundedMapViewBoundingBox = mapViewBoundingBox.boundWgs84Coordinates();
-                            BoundingBox transformedBoundingBox = boundedMapViewBoundingBox.transform(projectionTransform);
-                            if (featureProjection.isUnit(Units.DEGREES)) {
-                                filterMaxLongitude = ProjectionConstants.WGS84_HALF_WORLD_LON_WIDTH;
-                            } else if (featureProjection.isUnit(Units.METRES)) {
-                                filterMaxLongitude = ProjectionConstants.WEB_MERCATOR_HALF_WORLD_WIDTH;
-                            }
-                            filterBoundingBox = transformedBoundingBox.expandCoordinates(filterMaxLongitude);
-                        }
-
-                        // Query for all rows
-                        try (FeatureCursor cursor = featureDao.query(columns)) {
-                            while (!task.isCancelled() && count.get() < maxFeatures
-                                    && cursor.moveToNext()) {
-                                try {
-                                    FeatureRow row = cursor.getRow();
-
-                                    // Process the feature row in the thread pool
-                                    FeatureRowProcessor processor = new FeatureRowProcessor(
-                                            task, database, featureDao, row, count, maxFeatures, editable, converter,
-                                            styleCache, filterBoundingBox, filterMaxLongitude, filter);
-                                    ThreadUtils.getInstance().runBackground(processor);
-                                } catch (Exception e) {
-                                    Log.e(GeoPackageMapFragment.class.getSimpleName(),
-                                            "Failed to display feature. database: " + database
-                                                    + ", feature table: " + features
-                                                    + ", row: " + cursor.getPosition(), e);
-                                }
-                            }
-
-                        }
-                    }
-                    indexer.close();
-
-                }
-            }
-        }
-    }
-
-    /**
-     * Process the feature index results
-     *
-     * @param task         The feature update task.
-     * @param indexResults The index results.
-     * @param database     The geoPackage to process features for.
-     * @param featureDao   The feature data access object.
-     * @param converter    Convert the features shapes to those that can go on a google map.
-     * @param styleCache   The style cache.
-     * @param count        Keeps track of how many features we have added to the map.
-     * @param maxFeatures  The maximum number of features we can add to the map.
-     * @param editable     True if the feature added to the map should look editable.
-     * @param filter       True if we should filter the features based on a bounding box.
-     */
-    private void processFeatureIndexResults(MapFeaturesUpdateTask task, FeatureIndexResults indexResults, String database, FeatureDao featureDao,
-                                            GoogleMapShapeConverter converter, StyleCache styleCache, AtomicInteger count, final int maxFeatures, final boolean editable,
-                                            boolean filter) {
-
-        try {
-            for (FeatureRow row : indexResults) {
-
-                if (task.isCancelled() || count.get() >= maxFeatures) {
-                    break;
-                }
-
-                try {
-
-                    // Process the feature row in the thread pool
-                    FeatureRowProcessor processor = new FeatureRowProcessor(
-                            task, database, featureDao, row, count, maxFeatures, editable, converter,
-                            styleCache, null, 0, filter);
-                    ThreadUtils.getInstance().runBackground(processor);
-
-                } catch (Exception e) {
-                    Log.e(GeoPackageMapFragment.class.getSimpleName(),
-                            "Failed to display feature. database: " + database
-                                    + ", feature table: " + featureDao.getTableName()
-                                    + ", row id: " + row.getId(), e);
-                }
-            }
-        } finally {
-            indexResults.close();
-        }
-    }
-
-    /**
-     * Single feature row processor
-     *
-     * @author osbornb
-     */
-    private class FeatureRowProcessor implements Runnable {
-
-        /**
-         * Map update task
-         */
-        private final MapFeaturesUpdateTask task;
-
-        /**
-         * Database
-         */
-        private final String database;
-
-        /**
-         * Feature DAO
-         */
-        private final FeatureDao featureDao;
-
-        /**
-         * Feature row
-         */
-        private final FeatureRow row;
-
-        /**
-         * Total feature count
-         */
-        private final AtomicInteger count;
-
-        /**
-         * Total max features
-         */
-        private final int maxFeatures;
-
-        /**
-         * Editable shape flag
-         */
-        private final boolean editable;
-
-        /**
-         * Shape converter
-         */
-        private final GoogleMapShapeConverter converter;
-
-        /**
-         * Style Cache
-         */
-        private final StyleCache styleCache;
-
-        /**
-         * Filter bounding box
-         */
-        private final BoundingBox filterBoundingBox;
-
-        /**
-         * Max projection longitude
-         */
-        private final double maxLongitude;
-
-        /**
-         * Filter flag
-         */
-        private final boolean filter;
-
-        /**
-         * Constructor
-         *
-         * @param task              The update task.
-         * @param database          The name of the geopackage the features belong too.
-         * @param featureDao        The feature data access object.
-         * @param row               The row to process.
-         * @param count             The current total count of features.
-         * @param maxFeatures       The maximum features to display on the map.
-         * @param editable          True if the feature should look editable on the map.
-         * @param converter         Converts the feature's shape to one to use on the map.
-         * @param styleCache        The style cache.
-         * @param filterBoundingBox The bounding box to use for filtering.
-         * @param maxLongitude      The maximum longitude.
-         * @param filter            True if we should filter using the passed in bounding box.
-         */
-        public FeatureRowProcessor(MapFeaturesUpdateTask task, String database, FeatureDao featureDao,
-                                   FeatureRow row, AtomicInteger count, int maxFeatures,
-                                   boolean editable, GoogleMapShapeConverter converter, StyleCache styleCache,
-                                   BoundingBox filterBoundingBox, double maxLongitude, boolean filter) {
-            this.task = task;
-            this.database = database;
-            this.featureDao = featureDao;
-            this.row = row;
-            this.count = count;
-            this.maxFeatures = maxFeatures;
-            this.editable = editable;
-            this.converter = converter;
-            this.styleCache = styleCache;
-            this.filterBoundingBox = filterBoundingBox;
-            this.maxLongitude = maxLongitude;
-            this.filter = filter;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public void run() {
-            processFeatureRow(task, database, featureDao, converter, styleCache, row, count, maxFeatures,
-                    editable, filterBoundingBox, maxLongitude, filter);
-        }
-
-    }
-
-    /**
-     * Process the feature row
-     *
-     * @param task         The map update task.
-     * @param database     The geopackage name the feature row belongs too.
-     * @param featureDao   The feature data access object.
-     * @param converter    Converts the feature shape to one that can be used on a google map.
-     * @param styleCache   The style cache.
-     * @param row          The row to process.
-     * @param count        The current feature count displayed on map.
-     * @param maxFeatures  The maximum features to display on the map.
-     * @param editable     True if the feature should look editable on the map.
-     * @param boundingBox  The bounding box to use to filter features.
-     * @param maxLongitude The maximum longitude.
-     * @param filter       True if we should filer using the bounding box.
-     */
-    private void processFeatureRow(MapFeaturesUpdateTask task, String database, FeatureDao featureDao,
-                                   GoogleMapShapeConverter converter, StyleCache styleCache, FeatureRow row, AtomicInteger count,
-                                   int maxFeatures, boolean editable, BoundingBox boundingBox, double maxLongitude,
-                                   boolean filter) {
-
-        boolean exists;
-        synchronized (featureShapes) {
-            exists = featureShapes.exists(row.getId(), database, featureDao.getTableName());
-        }
-
-        if (!exists) {
-
-            try {
-                GeoPackageGeometryData geometryData = row.getGeometry();
-                if (geometryData != null && !geometryData.isEmpty()) {
-
-                    final Geometry geometry = geometryData.getGeometry();
-
-                    if (geometry != null) {
-
-                        boolean passesFilter = true;
-
-                        if (filter && boundingBox != null) {
-                            GeometryEnvelope envelope = geometryData.getEnvelope();
-                            if (envelope == null) {
-                                envelope = GeometryEnvelopeBuilder.buildEnvelope(geometry);
-                            }
-                            if (envelope != null) {
-                                if (geometry.getGeometryType() == GeometryType.POINT) {
-                                    mil.nga.sf.Point point = (mil.nga.sf.Point) geometry;
-                                    passesFilter = TileBoundingBoxUtils.isPointInBoundingBox(point, boundingBox, maxLongitude);
-                                } else {
-                                    BoundingBox geometryBoundingBox = new BoundingBox(envelope);
-                                    passesFilter = TileBoundingBoxUtils.overlap(boundingBox, geometryBoundingBox, maxLongitude) != null;
-                                }
-                            }
-                        }
-
-                        if (passesFilter && count.getAndIncrement() < maxFeatures) {
-                            final long featureId = row.getId();
-                            final GoogleMapShape shape = converter.toShape(geometry);
-                            updateFeaturesBoundingBox(shape);
-                            prepareShapeOptions(shape, styleCache, row, editable, true);
-                            task.addToMap(featureId, database, featureDao.getTableName(), shape);
-                        }
-                    }
-                }
-            } catch (Exception e) {
-                new Handler(Looper.getMainLooper()).post(() -> {
-                    Toast toast = Toast.makeText(getContext(), "Error loading geometry", Toast.LENGTH_SHORT);
-                    toast.show();
-                });
-            }
-        }
-    }
-
-    /**
-     * Update the features bounding box with the shape
-     *
-     * @param shape The shape to use to expand the features bounding box.
-     */
-    private void updateFeaturesBoundingBox(GoogleMapShape shape) {
-        try {
-            featuresBoundingBoxLock.lock();
-            if (model.getFeaturesBoundingBox() != null) {
-                shape.expandBoundingBox(model.getFeaturesBoundingBox());
-            } else {
-                model.setFeaturesBoundingBox(shape.boundingBox());
-            }
-        } finally {
-            featuresBoundingBoxLock.unlock();
-        }
-    }
-
-    /**
-     * Prepare the shape options
-     *
-     * @param shape      map shape
-     * @param styleCache style cache
-     * @param featureRow feature row
-     * @param editable   editable flag
-     * @param topLevel   top level flag
-     */
-    private void prepareShapeOptions(GoogleMapShape shape, StyleCache styleCache, FeatureRow featureRow, boolean editable,
-                                     boolean topLevel) {
-
-        FeatureStyle featureStyle = null;
-        if (styleCache != null) {
-            featureStyle = styleCache.getFeatureStyleExtension().getFeatureStyle(featureRow, shape.getGeometryType());
-        }
-
-        switch (shape.getShapeType()) {
-
-            case LAT_LNG:
-                LatLng latLng = (LatLng) shape.getShape();
-                MarkerOptions markerOptions = getMarkerOptions(styleCache, featureStyle, editable, topLevel);
-                markerOptions.position(latLng);
-                shape.setShape(markerOptions);
-                shape.setShapeType(GoogleMapShapeType.MARKER_OPTIONS);
-                break;
-
-            case POLYLINE_OPTIONS:
-                PolylineOptions polylineOptions = (PolylineOptions) shape
-                        .getShape();
-                setPolylineOptions(styleCache, featureStyle, editable, polylineOptions);
-                break;
-
-            case POLYGON_OPTIONS:
-                PolygonOptions polygonOptions = (PolygonOptions) shape.getShape();
-                setPolygonOptions(styleCache, featureStyle, editable, polygonOptions);
-                break;
-
-            case MULTI_LAT_LNG:
-                MultiLatLng multiLatLng = (MultiLatLng) shape.getShape();
-                MarkerOptions sharedMarkerOptions = getMarkerOptions(styleCache, featureStyle, editable,
-                        false);
-                multiLatLng.setMarkerOptions(sharedMarkerOptions);
-                break;
-
-            case MULTI_POLYLINE_OPTIONS:
-                MultiPolylineOptions multiPolylineOptions = (MultiPolylineOptions) shape
-                        .getShape();
-                PolylineOptions sharedPolylineOptions = new PolylineOptions();
-                setPolylineOptions(styleCache, featureStyle, editable, sharedPolylineOptions);
-                multiPolylineOptions.setOptions(sharedPolylineOptions);
-                break;
-
-            case MULTI_POLYGON_OPTIONS:
-                MultiPolygonOptions multiPolygonOptions = (MultiPolygonOptions) shape
-                        .getShape();
-                PolygonOptions sharedPolygonOptions = new PolygonOptions();
-                setPolygonOptions(styleCache, featureStyle, editable, sharedPolygonOptions);
-                multiPolygonOptions.setOptions(sharedPolygonOptions);
-                break;
-
-            case COLLECTION:
-                @SuppressWarnings("unchecked")
-                List<GoogleMapShape> shapes = (List<GoogleMapShape>) shape
-                        .getShape();
-                for (int i = 0; i < shapes.size(); i++) {
-                    prepareShapeOptions(shapes.get(i), styleCache, featureRow, editable, false);
-                }
-                break;
-            default:
-        }
-
-    }
-
-    /**
-     * Get marker options
-     *
-     * @param styleCache   style cache
-     * @param featureStyle feature style
-     * @param editable     editable flag
-     * @param clickable    clickable flag
-     * @return marker options
-     */
-    private MarkerOptions getMarkerOptions(StyleCache styleCache, FeatureStyle featureStyle, boolean editable, boolean clickable) {
-        MarkerOptions markerOptions = new MarkerOptions();
-        if (editable) {
-            TypedValue typedValue = new TypedValue();
-            if (clickable) {
-                getResources().getValue(R.dimen.marker_edit_color, typedValue,
-                        true);
-            } else {
-                getResources().getValue(R.dimen.marker_edit_read_only_color,
-                        typedValue, true);
-            }
-            markerOptions.icon(BitmapDescriptorFactory.defaultMarker(typedValue
-                    .getFloat()));
-
-        } else if (styleCache == null || !styleCache.setFeatureStyle(markerOptions, featureStyle)) {
-
-            TypedValue typedValue = new TypedValue();
-            getResources().getValue(R.dimen.marker_color, typedValue, true);
-            markerOptions.icon(BitmapDescriptorFactory.defaultMarker(typedValue.getFloat()));
-        }
-
-        return markerOptions;
-    }
-
-    /**
-     * Set the Polyline Option attributes
-     *
-     * @param styleCache      style cache
-     * @param featureStyle    feature style
-     * @param editable        editable flag
-     * @param polylineOptions polyline options
-     */
-    private void setPolylineOptions(StyleCache styleCache, FeatureStyle featureStyle, boolean editable,
-                                    PolylineOptions polylineOptions) {
-        if (getActivity() != null) {
-            if (editable) {
-                polylineOptions.color(ContextCompat.getColor(getActivity(), R.color.polyline_edit_color));
-            } else if (styleCache == null || !styleCache.setFeatureStyle(polylineOptions, featureStyle)) {
-                polylineOptions.color(ContextCompat.getColor(getActivity(), R.color.polyline_color));
-            }
-        }
-    }
-
-    /**
-     * Set the Polygon Option attributes
-     *
-     * @param styleCache     style cache
-     * @param featureStyle   feature style
-     * @param editable       True if it should be displayed as editable.
-     * @param polygonOptions The polygon options to set.
-     */
-    private void setPolygonOptions(StyleCache styleCache, FeatureStyle featureStyle, boolean editable,
-                                   PolygonOptions polygonOptions) {
-        if (getActivity() != null) {
-            if (editable) {
-                polygonOptions.strokeColor(ContextCompat.getColor(getActivity(), R.color.polygon_edit_color));
-                polygonOptions.fillColor(ContextCompat.getColor(getActivity(), R.color.polygon_edit_fill_color));
-            } else if (styleCache == null || !styleCache.setFeatureStyle(polygonOptions, featureStyle)) {
-                polygonOptions.strokeColor(ContextCompat.getColor(getActivity(), R.color.polygon_color));
-                polygonOptions.fillColor(ContextCompat.getColor(getActivity(), R.color.polygon_fill_color));
-            }
-        }
-    }
-
-    /**
-     * Add editable shape
-     *
-     * @param featureId The id of the feature.
-     * @param shape     The shape to add.
-     * @return marker The google map marker to add.
-     */
-    private Marker addEditableShape(long featureId, GoogleMapShape shape) {
-
-        Marker marker;
-
-        if (shape.getShapeType() == GoogleMapShapeType.MARKER) {
-            marker = (Marker) shape.getShape();
-        } else {
-            marker = getMarker(shape);
-            if (marker != null) {
-                editFeatureObjects.put(marker.getId(), shape);
-            }
-        }
-
-        if (marker != null) {
-            editFeatureIds.put(marker.getId(), featureId);
-        }
-
-        return marker;
-    }
-
-    /**
-     * Add marker shape
-     *
-     * @param featureId The id of the feature.
-     * @param database  The name of the geopackage the feature belongs to.
-     * @param tableName The name of the layer the feature belongs to.
-     * @param shape     The shape to add.
-     */
-    private void addMarkerShape(long featureId, String database, String tableName, GoogleMapShape shape) {
-
-        if (shape.getShapeType() == GoogleMapShapeType.MARKER) {
-            Marker marker = (Marker) shape.getShape();
-            MarkerFeature markerFeature = new MarkerFeature(featureId, database, tableName);
-            markerIds.put(marker.getId(), markerFeature);
-        }
-    }
-
-    /**
-     * Get the first marker of the shape or create one at the location
-     *
-     * @param shape The shape to get the marker for.
-     * @return The marker to add to the map.
-     */
-    private Marker getMarker(GoogleMapShape shape) {
-
-        Marker marker = null;
-
-        switch (shape.getShapeType()) {
-
-            case MARKER:
-                Marker shapeMarker = (Marker) shape.getShape();
-                marker = createEditMarker(shapeMarker.getPosition());
-                break;
-
-            case POLYLINE:
-                Polyline polyline = (Polyline) shape.getShape();
-                LatLng polylinePoint = polyline.getPoints().get(0);
-                marker = createEditMarker(polylinePoint);
-                break;
-
-            case POLYGON:
-                Polygon polygon = (Polygon) shape.getShape();
-                LatLng polygonPoint = polygon.getPoints().get(0);
-                marker = createEditMarker(polygonPoint);
-                break;
-
-            case MULTI_MARKER:
-                MultiMarker multiMarker = (MultiMarker) shape.getShape();
-                marker = createEditMarker(multiMarker.getMarkers().get(0)
-                        .getPosition());
-                break;
-
-            case MULTI_POLYLINE:
-                MultiPolyline multiPolyline = (MultiPolyline) shape.getShape();
-                LatLng multiPolylinePoint = multiPolyline.getPolylines().get(0)
-                        .getPoints().get(0);
-                marker = createEditMarker(multiPolylinePoint);
-                break;
-
-            case MULTI_POLYGON:
-                MultiPolygon multiPolygon = (MultiPolygon) shape.getShape();
-                LatLng multiPolygonPoint = multiPolygon.getPolygons().get(0)
-                        .getPoints().get(0);
-                marker = createEditMarker(multiPolygonPoint);
-                break;
-
-            case COLLECTION:
-                @SuppressWarnings("unchecked")
-                List<GoogleMapShape> shapes = (List<GoogleMapShape>) shape
-                        .getShape();
-                for (GoogleMapShape listShape : shapes) {
-                    marker = getMarker(listShape);
-                    if (marker != null) {
-                        break;
-                    }
-                }
-                break;
-            default:
-        }
-
-        return marker;
-    }
-
-    /**
-     * Create an edit marker to edit polylines and polygons
-     *
-     * @param latLng The latitude and longitude of the markers location.
-     * @return The marker to add to the map.
-     */
-    private Marker createEditMarker(LatLng latLng) {
-        MarkerOptions markerOptions = new MarkerOptions();
-        markerOptions.position(latLng);
-        markerOptions.icon(BitmapDescriptorFactory
-                .fromResource(R.drawable.ic_shape_edit));
-        TypedValue typedValueWidth = new TypedValue();
-        getResources().getValue(R.dimen.shape_edit_icon_anchor_width,
-                typedValueWidth, true);
-        TypedValue typedValueHeight = new TypedValue();
-        getResources().getValue(R.dimen.shape_edit_icon_anchor_height,
-                typedValueHeight, true);
-        markerOptions.anchor(typedValueWidth.getFloat(),
-                typedValueHeight.getFloat());
-        return map.addMarker(markerOptions);
     }
 
     /**
@@ -4384,7 +3537,7 @@ public class GeoPackageMapFragment extends Fragment implements
     private void displayFeatureTiles(GeoPackageFeatureOverlayTable featureOverlayTable) {
 
         GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(featureOverlayTable.getDatabase());
-        Map<String, FeatureDao> daos = featureDaos.get(featureOverlayTable.getDatabase());
+        Map<String, FeatureDao> daos = model.getFeatureDaos().get(featureOverlayTable.getDatabase());
         if (daos != null && getActivity() != null) {
             FeatureDao featureDao = daos.get(featureOverlayTable.getFeatureTable());
 
@@ -4899,7 +4052,7 @@ public class GeoPackageMapFragment extends Fragment implements
     @Override
     public void onMapClick(@NonNull LatLng point) {
 
-        if (!editFeaturesMode) {
+        if (!model.isEditFeaturesMode()) {
 
             StringBuilder clickMessage = new StringBuilder();
 
@@ -4931,7 +4084,7 @@ public class GeoPackageMapFragment extends Fragment implements
                     for (GeoPackageTable features : database.getFeatures()) {
 
                         GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(database.getDatabase());
-                        Map<String, FeatureDao> databaseFeatureDaos = featureDaos.get(database.getDatabase());
+                        Map<String, FeatureDao> databaseFeatureDaos = model.getFeatureDaos().get(database.getDatabase());
 
                         if (geoPackage != null && databaseFeatureDaos != null) {
 
@@ -5048,7 +4201,7 @@ public class GeoPackageMapFragment extends Fragment implements
 
         String markerId = marker.getId();
 
-        if (editFeaturesMode) {
+        if (model.isEditFeaturesMode()) {
 
             // Handle clicks to edit contents of an existing feature
             if (editFeatureShape != null && editFeatureShape.contains(markerId)) {
@@ -5057,7 +4210,7 @@ public class GeoPackageMapFragment extends Fragment implements
             }
 
             // Handle clicks on an existing feature in edit mode
-            Long featureId = editFeatureIds.get(markerId);
+            Long featureId = model.getEditFeatureIds().get(markerId);
             if (featureId != null) {
                 editExistingFeatureClick(marker, featureId);
                 return true;
@@ -5079,7 +4232,7 @@ public class GeoPackageMapFragment extends Fragment implements
 
         } else {
             // Handle clicks on point markers
-            MarkerFeature markerFeature = markerIds.get(markerId);
+            MarkerFeature markerFeature = model.getMarkerIds().get(markerId);
             if (markerFeature != null) {
                 infoFeatureClick(markerFeature);
                 return true;
@@ -5214,9 +4367,9 @@ public class GeoPackageMapFragment extends Fragment implements
      * @param featureId The id of the feature being edited.
      */
     private void editExistingFeatureClick(final Marker marker, long featureId) {
-        final GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(editFeaturesDatabase);
+        final GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(model.getEditFeaturesDatabase());
         final FeatureDao featureDao = geoPackage
-                .getFeatureDao(editFeaturesTable);
+                .getFeatureDao(model.getEditFeaturesTable());
 
         final FeatureRow featureRow = featureDao.queryForIdRow(featureId);
 
@@ -5376,23 +4529,23 @@ public class GeoPackageMapFragment extends Fragment implements
                     .setMessage(
                             getString(R.string.edit_features_delete_label) + " "
                                     + geometryType.getName() + " from "
-                                    + editFeaturesDatabase + " - "
-                                    + editFeaturesTable + " (lat="
+                                    + model.getEditFeaturesDatabase() + " - "
+                                    + model.getEditFeaturesTable() + " (lat="
                                     + position.latitude + ", lon="
                                     + position.longitude + ") ?")
                     .setPositiveButton(
                             getString(R.string.edit_features_delete_label),
 
                             (dialog, which) -> {
-                                GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(editFeaturesDatabase);
+                                GeoPackage geoPackage = geoPackageViewModel.getGeoPackage(model.getEditFeaturesDatabase());
                                 try {
 
                                     FeatureDao featureDao = geoPackage
-                                            .getFeatureDao(editFeaturesTable);
+                                            .getFeatureDao(model.getEditFeaturesTable());
                                     featureDao.delete(featureRow);
                                     marker.remove();
-                                    editFeatureIds.remove(marker.getId());
-                                    GoogleMapShape featureObject = editFeatureObjects
+                                    model.getEditFeatureIds().remove(marker.getId());
+                                    GoogleMapShape featureObject = model.getEditFeatureObjects()
                                             .remove(marker.getId());
                                     if (featureObject != null) {
                                         featureObject.remove();
